refactor(web): extract useBookmarks hook
Move the bookmark domain out of App: state, initial fetch, snapshot
sync as sessions change, legacy migration, and optimistic toggles.
The hook takes the current sessions and exposes bookmarkedSessions,
isSessionBookmarked, and the toggle actions. bookmarkedSidebarSessionIds
stays in App (it depends on sidebar composition) and now uses the
hook's isSessionBookmarked. Fourth step of the App.tsx decomposition.
